+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #1349 Add e2e test for Hive Partitioned Data
 - #1396 Create tables from other RDBMS
 - #1427 Support for CONCAT alias operator
+- #1424 Add get physical plan with explain
 
 ## Improvements
 - #1325 Refactored CacheMachine.h and CacheMachine.cpp